Boss in BossRooms w/Health
At the current moment I found a way to bosses in special boss rooms with
health. I tested it out in Boss1 room subclass. It may not be the most
perfect way to do it but the general idea is there. I connected the
bosses health by changing the boss class so that it called upon the
health class in its constructor. I wanted to do some more but it's late
and will have to wait til tomorrow. Other changes: I made the boss class
abstract.